Move class WasRun to its own file
