+++ WasRun.py
@@ -1,0 +1,6 @@
+class WasRun:
+    def __init__(self, name):
+        self.was_run = None
+
+    def testMethod(self):
+        self.was_run = 1
