+++ script.py
@@ -1,9 +1,4 @@
-class WasRun:
-    def __init__(self, name):
-        self.was_run = None
-
-    def testMethod(self):
-        self.was_run = 1
+from WasRun import WasRun
 
 test = WasRun("testMethod")
 print(test.was_run)
